Fix bug in InsertSong: SQL error when checkbox not checked

+++ src/NeoTransposer/Controllers/InsertSong.php
@@ -35,7 +35,7 @@ class InsertSong
 			'highest_note' => $request->get('highest_note'),
 			'people_lowest_note' => $request->get('people_lowest_note'),
 			'people_highest_note' => $request->get('people_highest_note'),
-			'first_chord_is_tone' => str_replace('on', '1', $request->get('first_chord_is_key')),
+			'first_chord_is_tone' => (int) str_replace('on', '1', $request->get('first_chord_is_key')),
 			'slug' => $this->getSlug($request),
 		));
